+++ config/routes.rb
@@ -1,12 +1,10 @@
 ThreebymeServer::Application.routes.draw do
-
   get 's3_credentials/info'
 
   resources :s3_credentials
   resources :connections
   resources :users
   resources :version_compatibilities
-  resources :documentation, only: [:show]
 
   resources :connection, only: [] do
     post :set_visibility, on: :collection